apisample13.html

<!DOCTYPE HTML>
<title>Example with notifying iframe</title>
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>
<script>

var t = async_test("Containing document receives messages");
var start_received = false;
var result_received = false;
var completion_received = false;

var expected_messages = [
  t.step_func(function(message) {
    assert_equals("start", message.data.type);
    assert_own_property(message.data, "properties");
  }),

  t.step_func(function(message) {
    assert_equals("result", message.data.type);
    assert_equals(message.data.test.PASS, message.data.test.status);
  }),

  t.step_func(function(message) {
    assert_equals("complete", message.data.type);
    assert_equals(1, message.data.tests.length);
    assert_equals(message.data.tests[0].PASS, message.data.tests[0].status);
    assert_equals(message.data.status.OK, message.data.status.status);
    t.done();
  }),

  t.unreached_func("Too many messages received")
];

on_event(window, 'message', function(message) {
  var handler = expected_messages.shift();
  handler(message);
});

</script>
<iframe src="apisample6.html" style="display:none">
  <!-- apisample6 implements a file_is_test test. -->
</iframe>
